chore: fix sqa 2.0 type in 1.12.x (#27367)

+++ ingestion/tests/unit/topology/database/test_pinotdb.py
@@ -33,7 +33,7 @@ def _resolve(pinot_type: str) -> str:
 @pytest.mark.parametrize(
     "pinot_type, expected_om_type",
     [
-        ("double", "DOUBLE"),
+        ("double", "NUMERIC"),
         ("float", "FLOAT"),
         ("int", "BIGINT"),
         ("long", "BIGINT"),
@@ -53,4 +53,4 @@ def test_double_not_mapped_to_int():
     """Explicit regression test: Pinot DOUBLE must never resolve to INT."""
     result = _resolve("double")
     assert result != "INT", "Pinot DOUBLE is incorrectly mapped to INT"
-    assert result == "DOUBLE"
+    assert result == "NUMERIC"
